+++ packages/expo-asset/src/AssetSources.ts
@@ -92,30 +92,3 @@ export function resolveUri(uri: string): string {
   // `manifestBaseUrl` is always an absolute URL or `null`.
   return manifestBaseUrl ? new URL(uri, manifestBaseUrl).href : uri;
 }
-
-// A very cheap path canonicalization like path.join but without depending on a `path` polyfill.
-export function pathJoin(...paths: string[]): string {
-  // Start by simply combining paths, without worrying about ".." or "."
-  const combined = paths
-    .map((part, index) => {
-      if (index === 0) {
-        return part.trim().replace(/\/*$/, '');
-      }
-      return part.trim().replace(/(^\/*|\/*$)/g, '');
-    })
-    .filter((part) => part.length > 0)
-    .join('/')
-    .split('/');
-
-  // Handle ".." and "." in paths
-  const resolved: string[] = [];
-  for (const part of combined) {
-    if (part === '..') {
-      resolved.pop(); // Remove the last element from the result
-    } else if (part !== '.') {
-      resolved.push(part);
-    }
-  }
-
-  return resolved.join('/');
-}